Feature/fayzulloh

+++ src/modules/branches/branches.model.js
@@ -1,6 +1,19 @@
 const { error } = require('../../config/error.names');
 const { uniqRow } = require('../../lib/pg');
 
+async function readBranchModel() {
+    const result = await uniqRow('select * from branches');
+    if (!result.rows.length) {
+        return {
+            action: true,
+            status: 400,
+            error: error.BRANCH_NAME_NOT_FOUND,
+            message: 'Bunday nomli filial  topilmadi !',
+        };
+    }
+    return result.rows;
+}
+
 async function createBrachesModel(body) {
     const { branch_name } = body;
 
@@ -26,25 +39,14 @@ async function createBrachesModel(body) {
         };
     }
 }
-async function readBranchModel() {
-    const result = await uniqRow('select * from branches');
-    if (!result.rows.length) {
-        return {
-            action: true,
-            status: 400,
-            error: error.BRANCH_NAME_NOT_FOUND,
-            message: 'Bunday nomli filial  topilmadi !',
-        };
-    }
-    return result.rows;
-}
-async function updateBranchModel(params,body) {
+
+async function updateBranchModel(params, body) {
     const { branch_name } = body;
     const { branch_id } = params;
 
     const findedBranch = await uniqRow('select * from branches where branch_id = ($1)', branch_id);
-    
-    if(!findedBranch.rows.length){
+
+    if (!findedBranch.rows.length) {
         return {
             action: true,
             status: 404,
@@ -53,9 +55,11 @@ async function updateBranchModel(params,body) {
         };
     }
 
-    const findedName = await uniqRow('select * from branches where branch_name = $1', branch_name)
+    const branchName = branch_name ? branch_name : findedBranch.rows[0].branch_name
 
-    if(findedName.rows.length){
+    const findedName = await uniqRow('select * from branches where branch_name = $1', branch_name);
+
+    if (findedName.rows.length) {
         return {
             action: true,
             status: 404,
@@ -63,8 +67,8 @@ async function updateBranchModel(params,body) {
             message: `Kiritilgan filial allaqachon qo'shilgan, iltimos boshqa nom kiriting`,
         };
     }
-    
-    const result = await uniqRow('update branches set branch_name = $1 where branch_id= $2 returning *', branch_name,branch_id);
+
+    const result = await uniqRow('update branches set branch_name = $1 where branch_id= $2 returning *', branchName, branch_id);
     if (result.rows.length) {
         return result.rows[0];
     } else {
@@ -76,12 +80,13 @@ async function updateBranchModel(params,body) {
         };
     }
 }
+
 async function deleteBranchModel(body) {
     const { branch_id } = body;
 
     const findedBranch = await uniqRow('select * from branches where branch_id = ($1)', branch_id);
-    
-    if(!findedBranch.rows.length){
+
+    if (!findedBranch.rows.length) {
         return {
             action: true,
             status: 404,
@@ -90,14 +95,25 @@ async function deleteBranchModel(body) {
         };
     }
     
-    const result = await uniqRow('delete from branches where branch_id=($1) returning * ',branch_id);
+    const users = await uniqRow('select * from users where branch_id = $1', branch_id)
+
+    if (users.rows.length) {
+        return {
+            action: true,
+            status: 401,
+            error: error.BRANCH_DELETE_ERROR,
+            message: `O'chirmoqchi bo'lgan filial da bog'langan foydalanuvchi bor`,
+        };
+    }
+    
+    const result = await uniqRow('delete from branches where branch_id=($1) returning * ', branch_id);
 
     if (result.rows.length) {
         return result.rows[0];
     } else {
         return {
             action: true,
-            status: 400,
+            status: 401,
             error: error.BRANCH_DELETE_ERROR,
             message: 'Xatolik yuz berdi, qaytadan urunib ko`ring !',
         };
